Fix errors in Travis
Add extra error catching code

The client.on("error", ...) is what fixes the problem, by catching
errors that happen with the connection in the background. I suspect the
docker container is being stopped before the client has time to
disconnect, which was causing uncaught exceptions. We don't really care
about this (I think).