extend the MigrateDBConfig to pass the ssl property
This allows the usage of the ssl property, which is passed down to the pg
driver. The migrate function utilizes the full pg driver to connect to
the database and use the tls options.
Right now, it allows you to run migrations, no db creation.

This pr is used to avoid the error:

``` bash
error TS2345: Object literal may only specify known properties, and 'ssl' does not exist in type 'MigrateDBConfig'.
```